requirements.txt
asgiref==3.6.0
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.7
django-autoslug==1.9.8
django-cleanup==7.0.0
django-modeltranslation==0.18.9
django-rosetta==0.9.8
django-tinymce==3.5.0
easy-thumbnails==2.8.5
executing==1.2.0
idna==3.4
ipython==8.11.0
jedi==0.18.2
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
polib==1.2.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.14.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
text-unidecode==1.3
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
Unidecode==1.3.6
urllib3==1.26.14
wcwidth==0.2.6
